Fixed Circular import located in Clients package. This included folder restructuring
