+++ src/clients/provisioning_client.py
@@ -12,12 +12,12 @@
 # @link       https://github.com/Lockstep-Network/lockstep-sdk-python
 #
 
-from src.lockstep_api import LockstepApi
 from src.models.lockstep_response import LockstepResponse
 from src.models.provisioningmodel import ProvisioningModel
 from src.models.provisioningfinalizerequestmodel import ProvisioningFinalizeRequestModel
 
 class ProvisioningClient:
+    from src.lockstep_api.lockstep_api import LockstepApi
 
     def __init__(self, client: LockstepApi):
         self.client = client
