+++ src/clients/status_client.py
@@ -12,10 +12,10 @@
 # @link       https://github.com/Lockstep-Network/lockstep-sdk-python
 #
 
-from src.lockstep_api import LockstepApi
 from src.models.lockstep_response import LockstepResponse
 
 class StatusClient:
+    from src.lockstep_api.lockstep_api import LockstepApi
 
     def __init__(self, client: LockstepApi):
         self.client = client
