+++ src/models/notemodel.py
@@ -18,7 +18,7 @@ from dataclasses import dataclass
 """
 A note is a customizable text string that can be attached to various 
 account attributes within Lockstep. You can use notes for internal 
-communication, correspondence with clients, or personal reminders. The 
+communication, correspondence with clients_, or personal reminders. The 
 Note Model represents a note and a number of different metadata 
 attributes related to the creation, storage, and ownership of the note. 
 See [Extensibility](https://developer.lockstep.io/docs/extensibility)
